linux: Detect `sudo` in order to use `$SUDO_USER`
We now require that `$SUDO_USER` be defined, not only that the script is
being run as root.  This generally means that the script must be run
using the `sudo` command.  This requirement improves consistency and
provides a better guarantee that we can identify the logged-in user.

For a reason that I haven't yet identified, `logname` is not guaranteed
to return the current user name.  On one installation of Xubuntu,
running `$ sudo logname` (or `$ logname`) returns the following output:

logname: no login name